AddIncident: Improved location listening strategy
The previous implementation would often listen to only one provider. For
instance, if GPS were enabled, getBestProvider would return "gps" for
both the low and the high quality criteria. Now we explicitly attempt to
listen to both the GPS and the Network provider, this way if GPS is
enabled, but the user is inside, we still have a chance of getting a
fix from the network. This also ensures that if, say, GPS is enabled
after the activity started, we'll receive updates from it.

In addition, the last known location is retrieved and displayed while
waiting for a better fix.