Close #2247 unable to upload pics
Reports fails to upload saying network is slow even though it's not.
This happens because during the upload process, the app thinks there is
a photo and it tries to upload but that fails causing the API to return
the wrong json response.

Fixed this by putting a condition to determine, if the photo really
exists.