Close #2408 - Fix offline report not uploaded when there is internet
Reports stored to be sent later to the deployment, are not sent when
the devices gains internet connectivity. This feature works with reports
not checkins. Found that, the code only sends offline reports when
checkin is enabled. Had to make it the other way around.
Send offline reports not offline checkins.