Fix entered location from being reset when add a photo to a report
When submitting a report and attaching a photo to the report, after
going back to the report form, the location is updated to the
current location. Apparently, wrong value was being set when retrieving
the saved report. It was using the descript field instead. Fixed by
setting the right value for the location. Also the auto set is done only
when the location field is empty.